tests/requirements.txt
# Test framework dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0
python-dotenv==1.0.0

# API Gateway dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pydantic==2.5.2
pydantic[email]==2.5.2
alembic==1.12.1
httpx==0.25.2

# Natural Language Agent dependencies (optional - for advanced tests)
# openai==1.3.7
# neo4j==5.14.0  
# langchain==0.0.342

# Shared utils dependencies
kafka-python==2.0.2
redis==5.0.1
aiokafka==0.10.0

# Load testing dependencies
locust==2.17.0